Groups files in an MVC folder structure
